feat: allow different `secret.auto.tfvars` for different environments when running locally (#61)
Previously, we were placing `secret.auto.tfvars` in 
`modules/env_secrets`. This meant there was a single
`secret.auto.tfvars` for all environments, and we'd have to tweak the
file if we wanted to change the secrets (ie: targeting a different
Stripe Sandbox).

With this change, `secret.auto.tfvars` goes in
`envs/${environment}/env_secrets` so we can have 1 `secret.auto.tfvars`
file per environment and we avoid tweaking the file in place, which can
be dangerous considering these `secret.auto.tfvars` aren't in version
control (as they shouldn't be).

I've tried it on my machine both when the file exists (running locally)
and when the file doesn't exist (like when running from CI).